Rename file of codeMapNodeSelector correctly #3054
