+++ visualization/app/codeCharta/state/effects/updateVisibleTopLabels/updateVisibleTopLabels.effect.spec.ts
@@ -2,9 +2,9 @@ import { TestBed } from "@angular/core/testing"
 import { EffectsModule } from "../../angular-redux/effects/effects.module"
 import { ApplicationInitStatus } from "@angular/core"
 import { UpdateVisibleTopLabelsEffect } from "./updateVisibleTopLabels.effect"
-import { setAmountOfTopLabels } from "../../store/appSettings/amountOfTopLabels/amountOfTopLabels.actions"
-import { codeMapNodesSelector } from "../../selectors/accumulatedData/codeMapNodesSelector"
-import { Subject } from "rxjs"
+import {setAmountOfTopLabels} from "../../store/appSettings/amountOfTopLabels/amountOfTopLabels.actions"
+import {codeMapNodesSelector} from "../../selectors/accumulatedData/codeMapNodes.selector"
+import {Subject} from "rxjs"
 import { visibleFileStatesSelector } from "../../selectors/visibleFileStates.selector"
 import { Store } from "../../angular-redux/store"
 import { Store as PlainStoreUsedByEffects } from "../../store/store"
