+++ visualization/app/codeCharta/state/effects/updateVisibleTopLabels/updateVisibleTopLabels.effect.ts
@@ -2,9 +2,9 @@ import { Inject, Injectable } from "@angular/core"
 import { Store } from "../../angular-redux/store"
 import { createEffect } from "../../angular-redux/effects/createEffect"
 import { map, withLatestFrom } from "rxjs"
-import { visibleFileStatesSelector } from "../../selectors/visibleFileStates.selector"
-import { codeMapNodesSelector } from "../../selectors/accumulatedData/codeMapNodesSelector"
-import { setAmountOfTopLabels } from "../../store/appSettings/amountOfTopLabels/amountOfTopLabels.actions"
+import {visibleFileStatesSelector} from "../../selectors/visibleFileStates.selector"
+import {codeMapNodesSelector} from "../../selectors/accumulatedData/codeMapNodes.selector"
+import {setAmountOfTopLabels} from "../../store/appSettings/amountOfTopLabels/amountOfTopLabels.actions"
 import { getNumberOfTopLabels } from "./getNumberOfTopLabels"
 
 @Injectable()
